backend/requirements.txt
selenium==4.16.0
fastapi==0.109.0
uvicorn==0.25.0
python-dotenv==1.0.0
supabase>=2.0.0
gotrue==0.2.0
httpx==0.16.1
requests==2.25.1
webdriver_manager
urllib3 
scrapy>=2.11.0
scrapy-playwright>=0.0.34 